feat: Update naming according to code review
