--- components/bybit/actions/bybit-account-list-positions/bybit-account-list-positions.mjs
@@ -1,31 +1,0 @@
-import bybit from "../../bybit.app.mjs";
-
-export default {
-  name: "ByBit Account List Positions",
-  version: "0.0.1",
-  key: "bybit-account-list-positions",
-  description: "Get Positions List" +
-      " [reference](https://bybit-exchange.github.io/docs/futuresV2/linear/#t-myposition)",
-  props: {
-    bybit,
-    symbol: {
-      propDefinition: [
-        bybit,
-        "symbol",
-      ],
-    },
-  },
-  type: "action",
-  methods: {},
-  async run({ $ }) {
-    const API_METHOD = "GET";
-    const API_PATH = "/private/linear/position/list";
-    let returnValue = await this.bybit.makeRequest(API_METHOD, API_PATH, this);
-    if (returnValue.ret_code) {
-      throw new Error(returnValue.ret_msg);
-    } else {
-      $.export("$summary", "Position List Request Successfully");
-    }
-    return returnValue;
-  },
-};
