--- components/bybit/actions/bybit-market-get-tickers/bybit-market-get-tickers.mjs
@@ -1,31 +1,0 @@
-import bybit from "../../bybit.app.mjs";
-
-export default {
-  name: "ByBit Market Get Tickers",
-  version: "0.0.1",
-  key: "bybit-market-get-tickers",
-  description: "Get the latest information for symbol. [reference](https://bybit-exchange.github.io/docs/futuresV2/inverse/#t-latestsymbolinfo)",
-  props: {
-    bybit,
-    symbol: {
-      propDefinition: [
-        bybit,
-        "symbol",
-      ],
-      optional: true,
-    },
-  },
-  type: "action",
-  methods: {},
-  async run({ $ }) {
-    const API_METHOD = "GET";
-    const API_PATH = "/v2/public/tickers";
-    let returnValue = await this.bybit.makeRequest(API_METHOD, API_PATH, this);
-    if (returnValue.ret_code) {
-      throw new Error(returnValue.ret_msg);
-    } else {
-      $.export("$summary", "Get Ticker Information Successful");
-    }
-    return returnValue;
-  },
-};
