--- components/bybit/actions/bybit-market-order-book/bybit-market-order-book.mjs
@@ -1,30 +1,0 @@
-import bybit from "../../bybit.app.mjs";
-
-export default {
-  name: "ByBit Market Order Book",
-  version: "0.0.1",
-  key: "bybit-market-order-book",
-  description: "Get Order Book.[reference](https://bybit-exchange.github.io/docs/futuresV2/inverse/#t-orderbook)",
-  props: {
-    bybit,
-    symbol: {
-      propDefinition: [
-        bybit,
-        "symbol",
-      ],
-    },
-  },
-  type: "action",
-  methods: {},
-  async run({ $ }) {
-    const API_METHOD = "GET";
-    const API_PATH = "/v2/public/orderBook/L2";
-    let returnValue = await this.bybit.makeRequest(API_METHOD, API_PATH, this);
-    if (returnValue.ret_code) {
-      throw new Error(returnValue.ret_msg);
-    } else {
-      $.export("$summary", "Get Order Book Successful");
-    }
-    return returnValue;
-  },
-};
