--- components/bybit/actions/bybit-market-query-symbols/bybit-market-query-symbols.mjs
@@ -1,24 +1,0 @@
-import bybit from "../../bybit.app.mjs";
-
-export default {
-  name: "ByBit Market Query Symbols",
-  version: "0.0.1",
-  key: "bybit-market-query-symbols",
-  description: "Get symbol info.[reference](https://bybit-exchange.github.io/docs/futuresV2/inverse/#t-querysymbol)",
-  props: {
-    bybit,
-  },
-  type: "action",
-  methods: {},
-  async run({ $ }) {
-    const API_METHOD = "GET";
-    const API_PATH = "/v2/public/symbols";
-    let returnValue = await this.bybit.makeRequest(API_METHOD, API_PATH, this);
-    if (returnValue.ret_code) {
-      throw new Error(returnValue.ret_msg);
-    } else {
-      $.export("$summary", "Get Symbols Information Successful");
-    }
-    return returnValue;
-  },
-};
