--- components/bybit/actions/bybit-market-recent-trading-records/bybit-market-recent-trading-records.mjs
@@ -1,38 +1,0 @@
-import bybit from "../../bybit.app.mjs";
-
-export default {
-  name: "ByBit Market Recent Trading Records",
-  version: "0.0.1",
-  key: "bybit-market-recent-trading-records",
-  description: "Get recent trades.[reference](https://bybit-exchange.github.io/docs/futuresV2/linear/#t-latestsymbolinfo)",
-  props: {
-    bybit,
-    symbol: {
-      propDefinition: [
-        bybit,
-        "symbol",
-      ],
-    },
-    limit: {
-      propDefinition: [
-        bybit,
-        "limit",
-      ],
-      max: 1000,
-      description: "Limit for data size, max size is 1000. Default as showing 500 pieces of data",
-    },
-  },
-  type: "action",
-  methods: {},
-  async run({ $ }) {
-    const API_METHOD = "GET";
-    const API_PATH = "/public/linear/recent-trading-records";
-    let returnValue = await this.bybit.makeRequest(API_METHOD, API_PATH, this);
-    if (returnValue.ret_code) {
-      throw new Error(returnValue.ret_msg);
-    } else {
-      $.export("$summary", "Get recent trades Successful");
-    }
-    return returnValue;
-  },
-};
