+++ components/bybit/actions/linear-order-replace/linear-order-replace.mjs
@@ -1,0 +1,79 @@
+import bybit from "../../bybit.app.mjs";
+
+export default {
+  name: "ByBit Linear Order Replace",
+  version: "0.0.1",
+  key: "bybit-linear-order-replace",
+  description: "Replace active order." +
+      " [reference](https://bybit-exchange.github.io/docs/futuresV2/linear/#t-replaceactive)",
+  props: {
+    bybit,
+    symbol: {
+      propDefinition: [
+        bybit,
+        "symbol",
+      ],
+    },
+    order_id: {
+      propDefinition: [
+        bybit,
+        "order_id",
+      ],
+    },
+    order_link_id: {
+      propDefinition: [
+        bybit,
+        "order_link_id",
+      ],
+    },
+    p_r_qty: {
+      propDefinition: [
+        bybit,
+        "qty",
+      ],
+    },
+    p_r_price: {
+      propDefinition: [
+        bybit,
+        "price",
+      ],
+    },
+    take_profit: {
+      propDefinition: [
+        bybit,
+        "take_profit",
+      ],
+    },
+    stop_loss: {
+      propDefinition: [
+        bybit,
+        "stop_loss",
+      ],
+    },
+    tp_trigger_by: {
+      propDefinition: [
+        bybit,
+        "tp_trigger_by",
+      ],
+    },
+    sl_trigger_by: {
+      propDefinition: [
+        bybit,
+        "sl_trigger_by",
+      ],
+    },
+  },
+  type: "action",
+  methods: {},
+  async run({ $ }) {
+    const API_METHOD = "POST";
+    const API_PATH = "/private/linear/order/replace";
+    let returnValue = await this.bybit.makeRequest(API_METHOD, API_PATH, this);
+    if (returnValue.ret_code) {
+      throw new Error(returnValue.ret_msg);
+    } else {
+      $.export("$summary", "Replace Active Order Successful");
+    }
+    return returnValue;
+  },
+};
