+++ components/bybit/actions/linear-position-set-leverage/linear-position-set-leverage.mjs
@@ -1,0 +1,42 @@
+import bybit from "../../bybit.app.mjs";
+
+export default {
+  name: "ByBit Linear Position Set Leverage",
+  version: "0.0.1",
+  key: "bybit-linear-position-set-leverage",
+  description: "Set Leverage [reference](https://bybit-exchange.github.io/docs/futuresV2/linear/#t-setleverage)",
+  props: {
+    bybit,
+    symbol: {
+      propDefinition: [
+        bybit,
+        "symbol",
+      ],
+    },
+    buy_leverage: {
+      propDefinition: [
+        bybit,
+        "buy_leverage",
+      ],
+    },
+    sell_leverage: {
+      propDefinition: [
+        bybit,
+        "sell_leverage",
+      ],
+    },
+  },
+  type: "action",
+  methods: {},
+  async run({ $ }) {
+    const API_METHOD = "POST";
+    const API_PATH = "/private/linear/position/set-leverage";
+    let returnValue = await this.bybit.makeRequest(API_METHOD, API_PATH, this);
+    if (returnValue.ret_code) {
+      throw new Error(returnValue.ret_msg);
+    } else {
+      $.export("$summary", "Set Leverage Successful");
+    }
+    return returnValue;
+  },
+};
